+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- add support for disk encryption to the `compute-vm` module
 - new `datafusion` module
 - new `container-registry` module
 - new `artifact-registry` module